ENH: moved Consolidators to a separate module
