+++ src/bluesky/callbacks/core.py
@@ -16,6 +16,14 @@ from event_model import DocumentRouter
 
 from ..utils import ensure_uid
 
+MIMETYPE_LOOKUP = {
+    "hdf5": "application/x-hdf5",
+    "ADHDF5_SWMR_STREAM": "application/x-hdf5",
+    "AD_HDF5_SWMR_SLICE": "application/x-hdf5",
+    "AD_TIFF": "multipart/related;type=image/tiff",
+    "AD_HDF5_GERM": "application/x-hdf5",
+}
+
 logger = logging.getLogger(__name__)
